# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Oficina Mecânica",
    "version": "1.0.0",
    "description": "API backend para oficina mecânica com autenticação, ordens de serviço e gerenciamento de clientes, veículos, serviços e peças.",
    "contact": {
      "name": "Equipe Oficina",
      "email": "equipe@oficina.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/api",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Troca de óleo"
          },
          "slaMinutes": {
            "type": "integer",
            "example": 60
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 80.0
          }
        }
      },
      "Part": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Filtro de óleo"
          },
          "type": {
            "type": "string",
            "example": "Filtro"
          },
          "model": {
            "type": "string",
            "example": "Honda Civic"
          },
          "color": {
            "type": "string",
            "example": "Preto"
          },
          "quantity": {
            "type": "integer",
            "example": 10
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 35.0
          }
        }
      },
      "OrderService": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "externalId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEBIDA",
              "EM_DIAGNOSTICO",
              "AGUARDANDO_APROVACAO",
              "EM_EXECUCAO",
              "FINALIZADA",
              "ENTREGUE"
            ],
            "example": "RECEBIDA"
          },
          "description": {
            "type": "string",
            "example": "Troca de óleo e filtros"
          },
          "mechanicName": {
            "type": "string",
            "example": "João Silva"
          },
          "budgetValue": {
            "type": "number",
            "format": "float",
            "example": 150.0
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:00:00Z"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:00:00Z"
          },
          "clientPFId": {
            "type": "integer",
            "example": 1
          },
          "clientPJId": {
            "type": "integer",
            "example": null
          },
          "vehicleId": {
            "type": "integer",
            "example": 1
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderServiceService"
            }
          }
        }
      },
      "OrderServiceService": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "orderServiceId": {
            "type": "integer",
            "example": 1
          },
          "serviceId": {
            "type": "integer",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:00:00Z"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderServiceServicePart"
            }
          }
        }
      },
      "OrderServiceServicePart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "orderServiceServiceId": {
            "type": "integer",
            "example": 1
          },
          "partId": {
            "type": "integer",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "example": 2
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:00:00Z"
          },
          "part": {
            "$ref": "#/components/schemas/Part"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Recurso não encontrado."
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "⭐ Fase 02 — Requisitos Obrigatórios",
          "Ordens de Serviço"
        ],
        "summary": "Abertura de Ordem de Serviço (OS)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "vehicleId"
                ],
                "properties": {
                  "vehicleId": {
                    "type": "integer",
                    "example": 1
                  },
                  "clientPFId": {
                    "type": "integer",
                    "example": 1
                  },
                  "clientPJId": {
                    "type": "integer",
                    "example": null
                  },
                  "description": {
                    "type": "string",
                    "example": "Troca de óleo e filtros"
                  },
                  "mechanicName": {
                    "type": "string",
                    "example": "José Santos"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ordem criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderService"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "⭐ Fase 02 — Requisitos Obrigatórios",
          "Ordens de Serviço"
        ],
        "summary": "Listagem de Ordens de Serviço",
        "responses": {
          "200": {
            "description": "Lista de ordens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderService"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Buscar ordem por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordem encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderService"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "tags": [
          "⭐ Fase 02 — Requisitos Obrigatórios",
          "Ordens de Serviço"
        ],
        "summary": "Atualização de status da OS",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "RECEBIDA",
                      "EM_DIAGNOSTICO",
                      "AGUARDANDO_APROVACAO",
                      "EM_EXECUCAO",
                      "FINALIZADA",
                      "ENTREGUE"
                    ],
                    "example": "EM_EXECUCAO"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderService"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/service": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Adicionar serviço à ordem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "serviceId"
                ],
                "properties": {
                  "serviceId": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Serviço adicionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderServiceService"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/part": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Adicionar peça à ordem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orderServiceServiceId",
                  "partId",
                  "quantity"
                ],
                "properties": {
                  "orderServiceServiceId": {
                    "type": "integer",
                    "example": 1
                  },
                  "partId": {
                    "type": "integer",
                    "example": 1
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Peça adicionada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderServiceServicePart"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/progress": {
      "get": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Consultar progresso da ordem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Progresso da ordem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderService"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}